LayoutTests/imported/w3c/web-platform-tests/css/css-contain/content-visibility/content-visibility-048-expected.txt


PASS Fragment navigation with content-visibility; single text
FAIL Fragment navigation with content-visibility; range across blocks assert_equals: expected "text2" but got "unknown"
